feat(s3s-rfc2047): extract the RFC 2047 codec into a standalone crate
Move the utils::rfc2047 module (encode/decode plus 58 unit tests) out of
the s3s crate into s3s-rfc2047, which already holds the crates.io name
reservation. The s3s http layer now calls s3s_rfc2047::{encode, decode}
for x-amz-meta header handling.

The codec drops its now-stale allow(dead_code) and starts panic-free:
file-level deny of the panic-related restriction lints (sigv4 precedent),
with the two indexing slices rewritten as checked access
(strip_prefix/strip_suffix for the encoded-word delimiters, get for the
chunk slice). Behavior is unchanged.
