+++ crates/s3s/Cargo.toml
@@ -54,6 +54,7 @@ phf.workspace = true
 pin-project-lite.workspace = true
 quick-xml.workspace = true
 regex.workspace = true
+s3s-rfc2047 = { version = "0.16.0-alpha.1", path = "../s3s-rfc2047" }
 s3s-sigv2 = { version = "0.16.0-alpha.1", path = "../s3s-sigv2" }
 s3s-sigv4 = { version = "0.16.0-alpha.1", path = "../s3s-sigv4" }
 serde.workspace = true
