+++ crates/s3s/src/utils/mod.rs
@@ -5,7 +5,6 @@
 
 pub mod crypto;
 pub mod parser;
-pub mod rfc2047;
 
 pub mod format;
